+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Changed
+- Switch to just blacklist of sites to run on
+
+### Added
+- Browser action popup to add rule
+- Background script to determine activeness
 
 ## [1.1] - 2016-08-29
 ### Added